Fix geocoding for schools in Sofia, Bulgaria
Our geocoder lib uses Google’s API by default and Google seems to be
confused if we include the municipality’s name in the address, which
is somewhat understandable.

Moreover, the “Sofia (capital)” and “Sofia” for outside-the-capital
addresses also seems strange and confuses the geocoder, so we’re
changing that as well.