Update the Ruby version and some gems
This is to make the app install and run on a recent macOS (Catalina).
There were errors with older Ruby versions and errors compiling the
native extensions of some older gems (nokogiri).

I'm also removing therubyracer as it's no longer maintained. The
buildpack in Heroku will likely have a JS runtime already.